requirements.txt
appdirs
asgiref
black
Django
django-widget-tweaks
flake8
isort
mccabe
mypy-extensions
pathspec
Pillow
pycodestyle
pyflakes
pytz
regex
sqlparse
tomli
django-ckeditor
djangorestframework